requirements.txt
streamlit
pandas
numpy
scikit-learn
faiss-cpu
sentence-transformers
beautifulsoup4
requests
pypdf
python-dotenv
groq
